team-matching: CPModel & CPSolver 다중해

+++ src/main/java/com/seoultech/ecc/team/repository/TimeRepository.java
@@ -3,5 +3,5 @@ package com.seoultech.ecc.team.repository;
 import com.seoultech.ecc.team.datamodel.TimeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TimeRepository extends JpaRepository<TimeEntity, Long> {
+public interface TimeRepository extends JpaRepository<TimeEntity, Integer> {
 }
